schoolmanagmentapp
The app is powered by the school

The app is administered by the school administration to schedule the activities of the teaching staff according to their class and gives parents access to the app to access their children's progress reports. It is a complete solution to monitor a child's progress directly and continuously.